luggages changes, my trips items views are working but we dont know why

+++ TravelSombrero-master/app/src/main/java/com/art/travelsombrero/LuggageDataModel.kt
@@ -1,5 +1,5 @@
 package com.art.travelsombrero
 
-class LuggageDataModel(val object_name: String, val amount: Long) {
+class LuggageDataModel(val object_name: String, val amount: Int) {
     constructor() : this("",0)
 }
